Decouples models (on disk) from repositories
  - Moves all files from src/models to src/repositories.
  - Creates models/StakeRequest.ts
  - Updates tests accordingly.
