+++ src/models/StakeRequest.ts
@@ -1,0 +1,34 @@
+// Copyright 2019 OpenST Ltd.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//    http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// ----------------------------------------------------------------------------
+
+/* eslint-disable class-methods-use-this */
+
+import BigNumber from 'bignumber.js';
+
+export default interface StakeRequest {
+  stakeRequestHash: string;
+  messageHash?: string;
+  amount: BigNumber;
+  beneficiary: string;
+  gasPrice: BigNumber;
+  gasLimit: BigNumber;
+  nonce: BigNumber;
+  gateway: string;
+  stakerProxy: string;
+  createdAt: Date;
+  updatedAt: Date;
+// eslint-disable-next-line semi
+}
