+++ test/repositories/AuxiliaryChainRepository/update.test.ts
@@ -21,9 +21,9 @@ import BigNumber from 'bignumber.js';
 import {
   AuxiliaryChainAttributes,
   AuxiliaryChain,
-} from '../../../src/models/AuxiliaryChainRepository';
+} from '../../../src/repositories/AuxiliaryChainRepository';
 
-import Database from '../../../src/models/Database';
+import Database from '../../../src/repositories/Database';
 
 import Util from './util';
